+++ src/lib/bq/regions.ts
@@ -444,3 +444,29 @@ export async function getKotters(
 
   return results || null;
 }
+
+export async function searchRegionsByName(q: string): Promise<RegionInfo[]> {
+  // Normalize and guard against overly-broad queries.
+  const term = (q || "").trim();
+  if (term.length < 2) return [];
+
+  // Escape single quotes to prevent SQL injection via LIKE.
+  const escapedTerm = term.replace(/'/g, "''").toLowerCase();
+
+  // Simple contains search; ordering is alphabetical for predictability.
+  const query = `
+    SELECT
+      region_id,
+      region_name,
+      logo_url,
+      is_active
+    FROM pv_regions
+    WHERE region_name IS NOT NULL
+      AND LOWER(region_name) LIKE '%${escapedTerm}%'
+    ORDER BY region_name
+    LIMIT 50
+  `;
+
+  const results = await queryBigQuery<RegionInfo>(query);
+  return results ?? [];
+}
